apis/events: change SessionId type from `int64` to `int`
This change doesn't affect the possible values that can be assigned to
`SessionId` since ChiChi is always compiled for a 64-bit architecture.

However, it does alter the requirement that when an event is converted
into a `map[string]any` value, it must be of type `int` instead of
`int64`, as it needs to match one of the internal data types.

This change prevents a panic that occurs during the marshaling of values
when they are sent to a transformation function.

+++ apis/events/event.go
@@ -135,7 +135,7 @@ type eventContext struct {
 		Height  int     `json:"height,omitempty"`
 		Density float64 `json:"density,omitempty"`
 	} `json:"screen,omitempty"`
-	SessionId    int64          `json:"sessionId,omitempty"`
+	SessionId    int            `json:"sessionId,omitempty"`
 	SessionStart bool           `json:"sessionStart,omitempty"`
 	GroupId      string         `json:"groupId,omitempty"`
 	Timezone     string         `json:"timezone,omitempty"`
